backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
qdrant-client
google-generativeai
langgraph
websockets
python-dotenv
httpx
numpy
scikit-learn
networkx